* moves filter for recents logic to RecentObjectsManager class * moves RecentObjectsManager class to services folder
